Add options to AuthenticatorButton
